Set stoploss on Apollo11 to -16%
Effectively, the tank check was not working as its author intended, ie,
if it's tanking set a stoploss of -10%.
What it was doing instead was setting the stop loss at -10% of the
current value, ie, if the current profit was -6%, the stoploss would be
set to -16%. This, once 30 minutes had passed.

This change just set's the stop loss at -16% from the start and removes
that tank check.